pyproject.toml
[tool.poetry]
name = "k-hop"
version = "0.2.0"
description = ""
authors = ["SC"]

[tool.poetry.dependencies]
python = "^3.8"
findspark = "^2.0.0"
pyspark = "^3.2.1"
argparse = "^1.4.0"
numpy = "^1.22.1"
graphframes-dev = "^0.7.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^22.3.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 79
include = '\/.*\.pyi?$'
